make runnable dir (#9016)
