+++ libs/langchain/langchain/schema/runnable/utils.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import asyncio
+from typing import Any, Coroutine, Union
+
+
+async def gated_coro(semaphore: asyncio.Semaphore, coro: Coroutine) -> Any:
+    async with semaphore:
+        return await coro
+
+
+async def gather_with_concurrency(n: Union[int, None], *coros: Coroutine) -> list:
+    if n is None:
+        return await asyncio.gather(*coros)
+
+    semaphore = asyncio.Semaphore(n)
+
+    return await asyncio.gather(*(gated_coro(semaphore, c) for c in coros))
